requirements.txt
Flask==3.1.3
openpyxl==3.1.5
GitPython==3.1.46
python-dotenv==1.0.0
PyYAML==6.0.3
pytest==8.4.0
